+++ src/Features/CSharpTest/Copilot/CSharpImplementNotImplementedExceptionFixProviderTests.cs
@@ -31,315 +31,6 @@ using VerifyCS = CSharpCodeFixVerifier<
 public sealed partial class CSharpImplementNotImplementedExceptionFixProviderTests
 {
     [Fact]
-    public async Task FixAll_ComplexScenarios_NestedThrowExpressions()
-    {
-        await new CustomCompositionCSharpTest
-        {
-            CodeFixTestBehaviors = CodeFixTestBehaviors.SkipLocalDiagnosticCheck,
-            TestCode = """
-            using System;
-            using System.Collections.Generic;
-            using System.Linq;
-            using System.Threading.Tasks;
-        
-            public class ComplexExceptionScenarios
-            {
-                public class Person
-                {
-                    public string Name { get; set; }
-                    public int Age { get; set; }
-                    public List<string> Skills { get; set; }
-                    public Person Supervisor { get; set; }
-                }
-
-                // Switch expression in lambda in a method
-                public void {|IDE3000:ProcessData|}(List<object> data)
-                {
-                    var result = data.Select(item => item switch 
-                    {
-                        string s => s.ToUpper(),
-                        int i => i.ToString(),
-                        DateTime d => d.ToShortDateString(),
-                        _ => {|IDE3000:throw new NotImplementedException("Unsupported data type")|} 
-                    });
-
-                    {|IDE3000:throw new NotImplementedException("Method not implemented");|}
-                }
-
-                // Object initializer with conditional expressions
-                public Person {|IDE3000:CreatePerson|}(string name, int age)
-                {
-                    return new Person
-                    {
-                        Name = name ?? {|IDE3000:throw new NotImplementedException("Name cannot be null")|},
-                        Age = age < 0 ? {|IDE3000:throw new NotImplementedException("Age must be positive")|} : age,
-                        Skills = new() { "C#", "F#" }
-                    };
-                }
-
-                // Local method with throw
-                public void {|IDE3000:ProcessWithLocalMethod|}(string input)
-                {
-                    string ParseInput(string text)
-                    {
-                        return text?.Length > 5 ? text : {|IDE3000:throw new NotImplementedException("Input too short")|};
-                    }
-
-                    {|IDE3000:throw new NotImplementedException("Method not fully implemented");|}
-                }
-
-                // Anonymous method with throw
-                public Func<int, int> {|IDE3000:GetCalculator|}(string operation)
-                {
-                    return operation switch
-                    {
-                        "square" => x => x * x,
-                        "double" => x => x * 2,
-                        _ => {|IDE3000:throw new NotImplementedException($"Operation {operation} not implemented")|} 
-                    };
-                }
-
-                // Nested delegated expressions
-                public void {|IDE3000:ProcessWithNestedDelegates|}()
-                {
-                    // Anonymous function that returns another function
-                    Func<int, Func<int, int>> createOperation = x => 
-                        y => x > 0 ? x + y : {|IDE3000:throw new NotImplementedException("Negative values not implemented")|};
-                
-                    {|IDE3000:throw new NotImplementedException("Method body not implemented");|}
-                }
-
-                // Async method with complex initialization
-                public async Task<Person> {|IDE3000:GetPersonAsync|}(int id)
-                {
-                    var supervisor = id > 100 
-                        ? new Person { Name = "Manager" } 
-                        : {|IDE3000:throw new NotImplementedException("Non-manager employees not implemented")|};
-                    
-                    {|IDE3000:throw new NotImplementedException("Async retrieval not implemented");|}
-                }
-            }
-            """,
-            FixedCode = """
-            using System;
-            using System.Collections.Generic;
-            using System.Linq;
-            using System.Threading.Tasks;
-        
-            public class ComplexExceptionScenarios
-            {
-                public class Person
-                {
-                    public string Name { get; set; }
-                    public int Age { get; set; }
-                    public List<string> Skills { get; set; }
-                    public Person Supervisor { get; set; }
-                }
-
-                // Switch expression in lambda in a method
-                public void ProcessData(List<object> data)
-                {
-                    var result = data.Select(item => item switch
-                    {
-                        string s => s.ToUpper(),
-                        int i => i.ToString(),
-                        DateTime d => d.ToShortDateString(),
-                        bool b => b.ToString().ToLower(),
-                        double d => d.ToString("F2"),
-                        List<object> list => $"List with {list.Count} items",
-                        _ => item?.ToString() ?? "null"
-                    }).ToList();
-
-                    foreach (var item in result)
-                    {
-                        Console.WriteLine($"Processed: {item}");
-                    }
-                }
-
-                // Object initializer with conditional expressions
-                public Person CreatePerson(string name, int age)
-                {
-                    return new Person
-                    {
-                        Name = name ?? throw new ArgumentNullException(nameof(name), "Name cannot be null"),
-                        Age = age < 0 ? throw new ArgumentException("Age must be positive", nameof(age)) : age,
-                        Skills = new() { "C#", "F#" }
-                    };
-                }
-
-                // Local method with throw
-                public void ProcessWithLocalMethod(string input)
-                {
-                    string ParseInput(string text)
-                    {
-                        return text?.Length > 5 ? text : throw new ArgumentException("Input must be longer than 5 characters", nameof(text));
-                    }
-
-                    if (string.IsNullOrEmpty(input))
-                    {
-                        throw new ArgumentNullException(nameof(input));
-                    }
-
-                    var result = ParseInput(input);
-                    Console.WriteLine($"Processed input: {result}");
-                }
-
-                // Anonymous method with throw
-                public Func<int, int> GetCalculator(string operation)
-                {
-                    return operation switch
-                    {
-                        "square" => x => x * x,
-                        "double" => x => x * 2,
-                        "triple" => x => x * 3,
-                        "increment" => x => x + 1,
-                        "decrement" => x => x - 1,
-                        _ => throw new ArgumentException($"Operation '{operation}' is not supported", nameof(operation))
-                    };
-                }
-
-                // Nested delegated expressions
-                public void ProcessWithNestedDelegates()
-                {
-                    // Anonymous function that returns another function
-                    Func<int, Func<int, int>> createOperation = x =>
-                        y => x > 0 ? x + y : throw new ArgumentException("First parameter must be positive", nameof(x));
-
-                    var operation = createOperation(10);
-                    var result = operation(20);
-                    Console.WriteLine($"Result of nested operation: {result}");
-                }
-
-                // Async method with complex initialization
-                public async Task<Person> GetPersonAsync(int id)
-                {
-                    var supervisor = id > 100
-                        ? new Person { Name = "Manager" }
-                        : new Person { Name = "Team Lead" };
-
-                    await Task.Delay(100); // Simulate database access
-
-                    return new Person
-                    {
-                        Name = $"Employee {id}",
-                        Age = 30 + id % 20,
-                        Skills = new List<string> { "C#", "SQL", "JavaScript" },
-                        Supervisor = supervisor
-                    };
-                }
-            }
-            """,
-            LanguageVersion = LanguageVersion.CSharp11,
-            ReferenceAssemblies = ReferenceAssemblies.Net.Net60,
-        }
-        .WithMockCopilotService(copilotService =>
-        {
-            copilotService.SetupFixAll = (Document document, ImmutableDictionary<SyntaxNode, ImmutableArray<ReferencedSymbol>> memberReferences, CancellationToken cancellationToken) =>
-            {
-                // Create a map of method/property implementations
-                var implementationMap = new Dictionary<string, string>
-                {
-                    ["ProcessData"] = """
-                public void ProcessData(List<object> data)
-                {
-                    var result = data.Select(item => item switch 
-                    {
-                        string s => s.ToUpper(),
-                        int i => i.ToString(),
-                        DateTime d => d.ToShortDateString(),
-                        bool b => b.ToString().ToLower(),
-                        double d => d.ToString("F2"),
-                        List<object> list => $"List with {list.Count} items",
-                        _ => item?.ToString() ?? "null"
-                    }).ToList();
-                    
-                    foreach (var item in result)
-                    {
-                        Console.WriteLine($"Processed: {item}");
-                    }
-                }
-                """,
-                    ["CreatePerson"] = """
-                public Person CreatePerson(string name, int age)
-                {
-                    return new Person
-                    {
-                        Name = name ?? throw new ArgumentNullException(nameof(name), "Name cannot be null"),
-                        Age = age < 0 ? throw new ArgumentException("Age must be positive", nameof(age)) : age,
-                        Skills = new() { "C#", "F#" }
-                    };
-                }
-                """,
-                    ["ProcessWithLocalMethod"] = """
-                public void ProcessWithLocalMethod(string input)
-                {
-                    string ParseInput(string text)
-                    {
-                        return text?.Length > 5 ? text : throw new ArgumentException("Input must be longer than 5 characters", nameof(text));
-                    }
-                    
-                    if (string.IsNullOrEmpty(input))
-                    {
-                        throw new ArgumentNullException(nameof(input));
-                    }
-                    
-                    var result = ParseInput(input);
-                    Console.WriteLine($"Processed input: {result}");
-                }
-                """,
-                    ["GetCalculator"] = """
-                public Func<int, int> GetCalculator(string operation)
-                {
-                    return operation switch
-                    {
-                        "square" => x => x * x,
-                        "double" => x => x * 2,
-                        "triple" => x => x * 3,
-                        "increment" => x => x + 1,
-                        "decrement" => x => x - 1,
-                        _ => throw new ArgumentException($"Operation '{operation}' is not supported", nameof(operation))
-                    };
-                }
-                """,
-                    ["ProcessWithNestedDelegates"] = """
-                public void ProcessWithNestedDelegates()
-                {
-                    // Anonymous function that returns another function
-                    Func<int, Func<int, int>> createOperation = x => 
-                        y => x > 0 ? x + y : throw new ArgumentException("First parameter must be positive", nameof(x));
-                    
-                    var operation = createOperation(10);
-                    var result = operation(20);
-                    Console.WriteLine($"Result of nested operation: {result}");
-                }
-                """,
-                    ["GetPersonAsync"] = """
-                public async Task<Person> GetPersonAsync(int id)
-                {
-                    var supervisor = id > 100 
-                        ? new Person { Name = "Manager" } 
-                        : new Person { Name = "Team Lead" };
-
-                    await Task.Delay(100); // Simulate database access
-
-                    return new Person
-                    {
-                        Name = $"Employee {id}",
-                        Age = 30 + id % 20,
-                        Skills = new List<string> { "C#", "SQL", "JavaScript" },
-                        Supervisor = supervisor
-                    };
-                }
-                """
-                };
-
-                return BuildResult(memberReferences, implementationMap);
-            };
-        })
-        .RunAsync();
-    }
-
-    [Fact]
     public async Task FixAll_ParseSuccessfully()
     {
         await new CustomCompositionCSharpTest
